저장하기 기능 추가 검색후 프로그램 종료시 chrome driver 가 남는 현상 발생

+++ SGI.py
@@ -10,22 +10,26 @@ import tkinter as tk
 import threading,queue
 
 class MainFrame(tk.Tk):
-    driver = ''
+    driver = None
     images = ''
     q = ''
+    save_active=True
     def __init__(self):
         tk.Tk.__init__(self)
         self.geometry("500x300+100+100")
         self.title("이미지 검색 툴")
-
         #저장위치 프레임
         saveFrame=tk.LabelFrame(self, text="저장위치설정", relief="solid", width="500", height="100", bd=1)
-        saveFrame.pack(side="top", fill="both", expand=False, padx=5, pady=2,)
+        saveFrame.pack(side="top", fill="both", expand=False, padx=5, pady=2)
 
         #위치 적히는 entry
         self.pathEntry = tk.Entry(saveFrame, bd=2, width=60)
-        self.pathEntry.pack(pady=5)
-        self.pathEntry.bind("<Button-1>", self.path_click)
+        self.pathEntry.insert(0, os.path.dirname(__file__)[:-6])
+        self.pathEntry.configure(state="readonly")
+        self.pathEntry.pack(side="left", anchor="center", pady=7,padx=10)
+
+        self.pathButton=tk.Button(saveFrame, command=self.path_click, text="...", width="3")
+        self.pathButton.pack(side="left")
         #검색하는 프레임
         searchFrame=tk.LabelFrame(self, text="검색 옵션", relief="solid", width="500", height="200", bd=1)
         searchFrame.pack(side="top", fill="both", expand=False, padx=5, pady=2)
@@ -79,8 +83,8 @@ class MainFrame(tk.Tk):
         stateFrame=tk.LabelFrame(self, text="상태", relief="solid", width="500", height="100", bd=1)
         stateFrame.pack(side="top", fill="both", expand=False, padx=5, pady=2, ipadx=2, ipady=2)
         #프로그레스 바
-        progressbar = tk.ttk.Progressbar(stateFrame, maximum=2000, length=480)
-        progressbar.grid(column=0, row=1, columnspan=20, pady=5, padx=5)
+        self.progressbar = tk.ttk.Progressbar(stateFrame, maximum=2000, length=480)
+        self.progressbar.grid(column=0, row=1, columnspan=20, pady=5, padx=5)
         #현재진행정도
         self.progressLabel=tk.Label(stateFrame,text="0")
         self.progressLabel.grid(column=0,row=0)
@@ -94,16 +98,20 @@ class MainFrame(tk.Tk):
         actionFrame=tk.Frame(self, relief="solid", width="500", height="50")
         actionFrame.pack(side="top", fill="both", expand=False, padx=4, pady=10)
         # 중지 버튼
-        self.stopButton = tk.Button(actionFrame, text="정지", width="8", state=tk.DISABLED)
+        self.stopButton = tk.Button(actionFrame, text="정지", width="8", state=tk.DISABLED, command=self.download_stop)
         self.stopButton.pack(side="right", padx=3)
         # 시작 버튼
-        self.startButton = tk.Button(actionFrame, text="다운로드", width="8")
+        self.startButton = tk.Button(actionFrame, text="다운로드", width="8", command=self.down_action)
         self.startButton.pack(side="right", padx=3, anchor="s")
+
     #저장위치 설정 클릭시 event 처리
-    def path_click(self,event):
+    def path_click(self):
         filePath = filedialog.askdirectory(initialdir="/", title='Please select a directory')
-        self.pathEntry.delete(0, "end")
-        self.pathEntry.insert(0, filePath)
+        self.pathEntry.configure(state=tk.NORMAL)
+        if not filePath =='':
+            self.pathEntry.delete(0, "end")
+            self.pathEntry.insert(0, filePath)
+        self.pathEntry.configure(state="readonly")
     #검색 속성 가져오기
     def search_stop_event(self):
         global driver
@@ -137,11 +145,12 @@ class MainFrame(tk.Tk):
                     "흰색": "white%2C", "회색": "ic:specific%2Cisc:gray%2C", "검정색": "ic:specific%2Cisc:black%2C", "갈색": "ic:specific%2Cisc:brown%2C"}
         #스크롤 이벤트
         options = webdriver.ChromeOptions()
-        options.add_argument('headless')
+        #options.add_argument('headless')
         options.add_argument('window-size=1920x1080')
         options.add_argument("disable-gpu")
+
         global driver
-        driver = webdriver.Chrome()
+        driver = webdriver.Chrome(options=options)
         pageUrl="http://www.google.co.kr/search?q="+name+"&tbm=isch&hl=ko&tbs="+color_dict[color]+time_dict[_time]+type_dict[type]+size_dict[size]
         result = pageUrl[len(pageUrl)-3:len(pageUrl)]
         if result == "%2C":
@@ -187,10 +196,49 @@ class MainFrame(tk.Tk):
         self.progressMaxLabel.configure(text=str(len(images)))
         self.search_startButton.configure(state=tk.NORMAL)
         self.search_stopButton.configure(state=tk.DISABLED)
+        self.progressbar.configure(maximum=len(images))
+        self.progressLabel.configure(text="0")
+        self.progressbar['value']=0
         messagebox.showinfo("검색완료","총파일수 : "+str(len(images)))
+    def down_action(self):
+        outPath = self.pathEntry.get()
+        keyName = self.search_Entry.get()
+        down_t=threading.Thread(target=MainFrame.download, args=(self,outPath,keyName), daemon=True)
+        down_t.start()
+    def download_stop(self):
+        global save_active
+        save_active=False
 
+    def download(self,outPath,keyName):
 
+        print("save btn action 실행")
+        q = queue.Queue()
+        # 버튼 중지하기 버튼으로 변경
+        self.startButton.configure(state=tk.DISABLED)
+        self.stopButton.configure(state=tk.NORMAL)
+        try:
+            global save_active
+            save_active = True
+            # 폴더 생성
+            if not os.path.isdir(outPath + "/" + keyName):  # 폴더 존재하지 않으면 생성
+                os.makedirs(outPath + "/" + keyName)
+            # download
+            for idx, image in enumerate(images):
+                if save_active == False:
+                    return
+                image.click()
+                driver.implicitly_wait(20)
 
+                imgUrl = driver.find_element_by_css_selector('.n3VNCb').get_attribute('src')
+                urllib.request.urlretrieve(imgUrl, outPath + "/" + keyName + "/" + str(idx + 1) + ".jpg")
+                self.progressLabel.configure(text=idx + 1)
+                self.progressbar['value'] = idx + 1
+                self.progressbar.update()
+        except:
+            return -1
+        finally:
+            self.startButton.configure(state=tk.NORMAL)
+            self.stopButton.configure(state=tk.DISABLED)
 
 
 
@@ -201,5 +249,9 @@ class MainFrame(tk.Tk):
 
 #메인루프
 if __name__ == "__main__":
-    app = MainFrame()
-    app.mainloop()
+    try:
+        app = MainFrame()
+        app.mainloop()
+        app.driver.quit()
+    except:
+        print("정상종료 실패")
